requirements.txt
fastapi==0.114.0
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.13
langchain-openai==0.1.15
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langdetect==1.0.9
langgraph==0.0.55
langserve==0.2.0
langsmith==0.1.83
pydantic==2.9.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.33.0
uvicorn==0.30.6
faiss-cpu==1.8.0
FlashRank==0.2.8
langchain_chroma==0.1.4
unstructured==0.15.13
rank-bm25==0.2.2
playwright
azure-ai-documentintelligence==1.0.0b4
streamlit_lottie
pymupdf
sse_starlette
# playwright install